requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==3.13.0
pyjwt==2.8.0
djangorestframework_simplejwt==5.3.0
psycopg2==2.9.9
